Split RPM::Spec specs almong module boundaries
