+++ spec/rpm/spec/requirements_spec.rb
@@ -1,0 +1,10 @@
+#
+# Licensed under the MIT license
+# Copyright (C) 2014 Red Hat, Inc.
+
+require 'polisher/rpm/spec'
+
+module Polisher::RPM
+  describe Spec do
+  end # describe Spec
+end # module Polisher::RPM
